cli/requirements.txt
mysql-connector-python
paho-mqtt